Fix image upload logic
