+++ app/globals.css
@@ -7,3 +7,7 @@ body {
 html {
   scroll-behavior: smooth;
 }
+
+.font-inter {
+  font-family: 'Inter', sans-serif;
+}
