+++ lib/lexical/lexicalConfig.ts
@@ -2,20 +2,18 @@
 
 import { InitialConfigType } from '@lexical/react/LexicalComposer';
 
-// Import all the necessary nodes
+// Lexical Nodes
 import { HeadingNode, QuoteNode } from '@lexical/rich-text';
 import { ListItemNode, ListNode } from '@lexical/list';
 import { CodeHighlightNode, CodeNode } from '@lexical/code';
 import { TableCellNode, TableNode, TableRowNode } from '@lexical/table';
 import { AutoLinkNode, LinkNode } from '@lexical/link';
-import { ImageNode } from '@/lib/lexical/ImageNode'; // Adjust path to your ImageNode file
+import { ImageNode } from '@/components/editor/ImageNode';
 
 export const editorConfig: InitialConfigType = {
   namespace: 'MyEditor',
   theme: {
-    // Paste your FULL theme object from the NewPostPage here
-    ltr: 'text-left',
-    rtl: 'text-right',
+    root: 'font-inter text-sm', // Apply Inter font to the editor content
     paragraph: 'mb-2',
     quote: 'border-l-4 border-gray-300 pl-4 italic my-4',
     heading: {
@@ -24,9 +22,10 @@ export const editorConfig: InitialConfigType = {
       h3: 'text-xl font-semibold mb-2',
     },
     list: {
-      ul: 'list-disc list-inside my-4 pl-4',
-      ol: 'list-decimal list-inside my-4 pl-4',
+      ul: 'list-disc list-outside my-4 pl-8',
+      ol: 'list-decimal list-outside my-4 pl-8',
     },
+    listitem: 'mb-2',
     link: 'text-teal-600 hover:underline',
     code: 'bg-zinc-700/50 text-sm p-4 my-4 block overflow-x-auto rounded-md',
     image: 'block my-4',
@@ -53,6 +52,6 @@ export const editorConfig: InitialConfigType = {
     TableRowNode,
     AutoLinkNode,
     LinkNode,
-    ImageNode, // Your custom image node
+    ImageNode,
   ],
 };
